[3] adding base integration test
