+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 *.iml
 .idea
 *.log
+*.pid
