+++ nginx/nginx.conf
@@ -1,7 +1,7 @@
 
 #user  nobody;
 worker_processes  1;
-daemon off;
+#daemon off;
 
 error_log  logs/error.log  notice;
