+++ tck/bin/run-tck.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+declare -r SOURCE=${BASH_SOURCE[0]}
+declare -r DIR=$(dirname ${SOURCE})
+
+${DIR}/../../node_modules/tape/bin/tape ${DIR}/../t/*.js
+
